% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allowed_types}
\alias{allowed_types}
\title{Allowed Question Types}
\format{
A character vector with 12 elements, including:
\describe{
\item{\code{"numerical"}}{Numerical questions}
\item{\code{"multichoice"}}{Multiple-choice questions}
\item{\code{"ordering"}}{Ordering questions}
\item{\code{"ordering<|>h"}}{Horizontal ordering questions}
\item{\code{"ordering<|>v"}}{Vertical ordering questions}
\item{\code{"ddwtos"}}{Drag-and-drop words into sentences}
\item{\code{"gapselect"}}{Gap select questions}
\item{\code{"matching"}}{Matching questions}
\item{\code{"essay"}}{Essay questions}
\item{\code{"truefalse"}}{True/False questions}
\item{\code{"shortanswer"}}{Short-answer questions}
\item{\code{"ddmarker"}}{Drag-and-drop markers}
}
}
\usage{
allowed_types
}
\description{
A vector of allowed question types for validation in the package.
}
\keyword{internal}
