% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_xml.R
\name{category_question}
\alias{category_question}
\title{Define the category of questions}
\usage{
category_question(category, questions)
}
\arguments{
\item{category}{A string, category name.}

\item{questions}{A string, formatted questions.}
}
\value{
A string.
}
\description{
Define the category of questions
}
\keyword{internal}
