% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{create_default_value_question_df}
\alias{create_default_value_question_df}
\title{Create a question data frame common to normal and extended questions with
default values.}
\usage{
create_default_value_question_df()
}
\description{
Create a question data frame common to normal and extended questions with
default values.
}
\keyword{internal}
