% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{create_question_csv}
\alias{create_question_csv}
\title{Create a question csv file}
\usage{
create_question_csv(file, sep = ",")
}
\arguments{
\item{file}{A string, name of a text file.}

\item{sep}{Column separator character.}
}
\value{
A string.
}
\description{
Creates an empty question csv file.
}
\examples{

file <- create_question_csv(file = tempfile(fileext = '.csv'))

}
\seealso{
Other support functions: 
\code{\link{create_question_data_frame}()},
\code{\link{read_question_csv}()},
\code{\link{vector_to_string}()}
}
\concept{support functions}
