% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{create_question_data_frame}
\alias{create_question_data_frame}
\title{Create a question data frame}
\usage{
create_question_data_frame()
}
\value{
A data frame.
}
\description{
Creates an empty question data frame.
}
\examples{

df <- create_question_data_frame()

}
\seealso{
Other support functions: 
\code{\link{create_question_csv}()},
\code{\link{read_question_csv}()},
\code{\link{vector_to_string}()}
}
\concept{support functions}
