% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{create_question_excel}
\alias{create_question_excel}
\title{Create a question Excel file}
\usage{
create_question_excel(file)
}
\arguments{
\item{file}{A string, name of a text file.}
}
\value{
A string.
}
\description{
Creates an empty question Excel file.
}
\examples{

file <- create_question_excel(file = tempfile(fileext = '.xlsx'))

}
\seealso{
Other support functions: 
\code{\link{create_question_csv}()},
\code{\link{create_question_data_frame}()},
\code{\link{read_question_csv}()},
\code{\link{read_question_excel}()},
\code{\link{vector_to_string}()}
}
\concept{support functions}
