% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extended_define_questions_from.R
\name{define_extended_questions_from_data_frame}
\alias{define_extended_questions_from_data_frame}
\title{Define extended questions from a data frame}
\usage{
define_extended_questions_from_data_frame(qc, df)
}
\description{
Define extended questions from a data frame
}
\keyword{internal}
