% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{define_question_csv}
\alias{define_question_csv}
\title{Define a question csv file}
\usage{
define_question_csv(file, sep = ",")
}
\arguments{
\item{file}{A string, name of a text file.}

\item{sep}{Column separator character.}
}
\value{
A string.
}
\description{
Defines an empty question csv file.
}
\examples{

file <- define_question_csv(file = tempfile(fileext = '.csv'))

}
\seealso{
Other support functions: 
\code{\link{define_question_data_frame}()},
\code{\link{vector_to_string}()}
}
\concept{support functions}
