% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{define_question_data_frame}
\alias{define_question_data_frame}
\title{Define a question data frame}
\usage{
define_question_data_frame()
}
\value{
A data frame.
}
\description{
Defines an empty question data frame.
}
\examples{

df <- define_question_data_frame()

}
\seealso{
Other support functions: 
\code{\link{define_question_csv}()},
\code{\link{vector_to_string}()}
}
\concept{support functions}
