% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_questions_extended.R
\name{define_questions_from_df}
\alias{define_questions_from_df}
\title{Define questions from a data frame}
\usage{
define_questions_from_df(qc, df)
}
\description{
Define questions from a data frame
}
\keyword{internal}
