% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_questions.R
\name{determine_orientation}
\alias{determine_orientation}
\title{Determine Orientation}
\usage{
determine_orientation(type)
}
\arguments{
\item{type}{A character string, typically "H", "h", or any other value.
\itemize{
\item "H" or "h": The orientation will be "h" (horizontal).
\item Any other value: The orientation will be "v" (vertical).
}}
}
\value{
A character string: either "h" for horizontal or "v" for vertical.
}
\description{
Determines the orientation based on the input type.
If the type is "H" or "h", the orientation will be horizontal ("h").
Otherwise, the orientation will be vertical ("v").
}
\keyword{internal}
