% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_question_type.R
\name{determine_question_type}
\alias{determine_question_type}
\title{Determine question type}
\usage{
determine_question_type(type, question, answer, a_values)
}
\arguments{
\item{type}{A string, suggested type for the question.}

\item{question}{A string, the main question text.}

\item{answer}{A string or vector, the correct answer(s) for the question.}

\item{a_values}{A vector, additional answers for the question.}
}
\value{
A string indicating the question type.
}
\description{
Determines the type of the question based on its content, provided answers,
and optional type parameter.
}
\keyword{internal}
