% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_questions.R
\name{filter_non_empty_answers}
\alias{filter_non_empty_answers}
\title{Filter non-empty answers}
\usage{
filter_non_empty_answers(...)
}
\arguments{
\item{...}{Strings, additional answers.}
}
\value{
A vector of non-empty answers.
}
\description{
Filters additional answers provided in \code{...} to ensure only non-empty answers
are included in the final output.
}
\keyword{internal}
