% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_xml.R
\name{format_questions}
\alias{format_questions}
\title{Format all questions in the data frame}
\usage{
format_questions(questions)
}
\arguments{
\item{questions}{A question data frame.}
}
\value{
A string.
}
\description{
Format all questions in the data frame
}
\keyword{internal}
