% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ddmarker.R
\name{generate_ddmarker}
\alias{generate_ddmarker}
\title{Generate \code{ddmarker} question}
\usage{
generate_ddmarker(image, image_alt, answer, fb_answer = "")
}
\arguments{
\item{image}{A string, optional, image file to include in the question.}

\item{image_alt}{A string, description of the image to include in the question.}

\item{answer}{A string vector.}

\item{fb_answer}{A string, answer feedback.}
}
\value{
A string.
}
\description{
Generate \code{ddmarker} question
}
\keyword{internal}
