% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_gapselect.R
\name{generate_gapselect}
\alias{generate_gapselect}
\title{Generate \code{gapselect} question}
\usage{
generate_gapselect(
  answer,
  rest,
  correct_feedback,
  incorrect_feedback,
  partially_correct_feedback
)
}
\arguments{
\item{answer}{A vector of strings.}

\item{rest}{A vector, rest of answers.}

\item{correct_feedback}{A string.}

\item{incorrect_feedback}{A string.}

\item{partially_correct_feedback}{A string.}
}
\value{
A string.
}
\description{
Generate \code{gapselect} question
}
\keyword{internal}
