% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_multichoice.R
\name{generate_multichoice}
\alias{generate_multichoice}
\title{Generate \code{multichoice} question}
\usage{
generate_multichoice(
  answer,
  a_values,
  correct_feedback,
  incorrect_feedback,
  fb_partially,
  fb_answer,
  fb_a_values,
  fraction
)
}
\arguments{
\item{answer}{A string.}

\item{a_values}{A vector, rest of answers.}

\item{correct_feedback}{A string.}

\item{incorrect_feedback}{A string.}

\item{fb_partially}{A string.}

\item{fb_answer}{A string, answer feedback.}

\item{fb_a_values}{A vector, rest of answer feedback.}

\item{fraction}{A number between 0 and 1.}
}
\value{
A string.
}
\description{
Generate \code{multichoice} question
}
\keyword{internal}
