% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_xml.R
\name{generate_name}
\alias{generate_name}
\title{generate \code{name} node}
\usage{
generate_name(first_question_number, type, orientation, question)
}
\arguments{
\item{first_question_number}{An integer, first number to compose the question
names.}

\item{type}{A string, question type (if needed).}

\item{orientation}{A string, 'h' or 'v'.}

\item{question}{A string, statement of the question.}
}
\value{
A string.
}
\description{
generate \code{name} node
}
\keyword{internal}
