% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ordering.R
\name{generate_ordering}
\alias{generate_ordering}
\title{Generate \code{ordering} question}
\usage{
generate_ordering(
  answer,
  n,
  rest,
  correct_feedback,
  partially_correct_feedback,
  incorrect_feedback,
  orientation
)
}
\arguments{
\item{answer}{A vector of strings.}

\item{n}{An integer, number or answers.}

\item{rest}{A vector, rest of answers.}

\item{correct_feedback}{A string.}

\item{partially_correct_feedback}{A string.}

\item{incorrect_feedback}{A string.}

\item{orientation}{A string, 'h' or 'v'.}
}
\value{
A string.
}
\description{
Generate \code{ordering} question
}
\keyword{internal}
