% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_questions.R
\name{generate_question_body}
\alias{generate_question_body}
\title{Generate the question body}
\usage{
generate_question_body(
  type,
  answer,
  rest,
  correct_feedback,
  incorrect_feedback,
  partially_correct_feedback,
  orientation
)
}
\arguments{
\item{type}{A string, the question type.}

\item{answer}{A string or vector, the correct answer(s) for the question.}

\item{rest}{A vector, additional answers for the question.}

\item{correct_feedback}{A string, feedback for correct answers.}

\item{incorrect_feedback}{A string, feedback for incorrect answers.}

\item{partially_correct_feedback}{A string, feedback for partially correct answers.}

\item{orientation}{A char, 'h' or 'v'.}
}
\value{
A string containing the question body in XML format.
}
\description{
Creates the body of the question based on its type and additional parameters
such as feedback and orientation.
}
\keyword{internal}
