% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_questions.R
\name{generate_question_name}
\alias{generate_question_name}
\title{generate question \code{name} node}
\usage{
generate_question_name(first_question_number, type, question)
}
\arguments{
\item{first_question_number}{An integer, first number to compose the question
names.}

\item{type}{A string, question type (if needed).}

\item{question}{A string, statement of the question.}
}
\value{
A string.
}
\description{
generate question \code{name} node
}
\keyword{internal}
