% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_xml.R
\name{generate_xml}
\alias{generate_xml}
\alias{generate_xml.question_category}
\title{Generar preguntas}
\usage{
generate_xml(qc, file)

\method{generate_xml}{question_category}(qc, file = NULL)
}
\arguments{
\item{file}{}
}
\description{
Generar preguntas
}
