% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_questions_extended.R
\name{get_detailed_type_names}
\alias{get_detailed_type_names}
\title{Get Detailed Type Names}
\usage{
get_detailed_type_names(df)
}
\arguments{
\item{df}{A data frame containing at least the following columns:
\itemize{
\item \code{type}: A character column representing the type of each question.
\item \code{answer}: A character column containing answers for each question.
\item Additional columns with a prefix "a_" representing associated answer fields.
}}
}
\value{
A character vector containing the updated \code{type} values.
}
\description{
Determines the \code{type} column of a data frame based on specific conditions and
the content of other columns in the data frame.
}
\details{
It checks if all the values in the \code{type} column are within the allowed set
(\verb{'', 'h', 'v', 'x'}) and updates each row based on the associated answers
and other fields.
}
\keyword{internal}
