% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extended_define_questions_from.R
\name{get_non_empty_fields_by_prefix}
\alias{get_non_empty_fields_by_prefix}
\title{Get non-empty values from fields with a specific prefix}
\usage{
get_non_empty_fields_by_prefix(df, i, prefix)
}
\arguments{
\item{df}{A dataframe containing the relevant columns.}

\item{i}{An integer representing the row index.}

\item{prefix}{A string representing the prefix of the column names (e.g., "a_", "fb_a_", "tag_").}
}
\value{
A vector with non-empty values from the fields with the given prefix in the specified row.
}
\description{
This function takes a dataframe, a row index, and a prefix, returning a vector
with the content of columns that start with the given prefix and are not empty.
}
\keyword{internal}
