% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_questions.R
\name{get_vector_answer}
\alias{get_vector_answer}
\title{Convert Answer String to Vector}
\usage{
get_vector_answer(answer)
}
\arguments{
\item{answer}{A character string representing the answer.}
}
\value{
A vector if the answer is successfully converted, or an empty string if the answer is \code{NULL}.
}
\description{
Converts an answer string into a vector using \code{string_to_vector}.
If the input is \code{NULL}, it returns an empty string.
}
\keyword{internal}
