% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{has_gaps}
\alias{has_gaps}
\title{Check if a String Contains Gaps}
\usage{
has_gaps(str)
}
\arguments{
\item{str}{\code{character}
A string to check.}
}
\value{
\code{logical}
\code{TRUE} if the string contains gaps, \code{FALSE} otherwise.
}
\description{
Checks if a string contains the specific substrings \code{"[[1]]"} and \code{"[[2]]"}.
}
\keyword{internal}
