% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{is_numeric}
\alias{is_numeric}
\title{Check if a String Represents a Numeric Value}
\usage{
is_numeric(str)
}
\arguments{
\item{str}{\code{character}
A string to check.}
}
\value{
\code{logical}
\code{TRUE} if the string is numeric, \code{FALSE} otherwise.
}
\description{
Verifies if a string can be converted to a numeric value.
}
\keyword{internal}
