% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{process_question_dataframe}
\alias{process_question_dataframe}
\title{Process Question DataFrame}
\usage{
process_question_dataframe(df)
}
\arguments{
\item{df}{A dataframe to process.}
}
\value{
A processed dataframe.
}
\description{
Processes a dataframe by converting columns to character type, handling NAs,
and renaming attributes to snake_case.
}
\keyword{internal}
