% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_category.R
\name{question_category}
\alias{question_category}
\title{\code{question_category} S3 class}
\usage{
question_category(
  category = "Default category",
  copyright = NULL,
  license = NULL,
  correct_feedback = "Correct.",
  partially_correct_feedback = "Partially correct.",
  incorrect_feedback = "Incorrect.",
  adapt_images = TRUE,
  width = 800,
  height = 600
)
}
\arguments{
\item{category}{A string, category name.}

\item{copyright}{A string, copyright text to be included in each question that
is defined.}

\item{license}{A string, license text to be included in each question that is
defined.}

\item{correct_feedback}{A string, feedback on correct answers to each question.}

\item{partially_correct_feedback}{A string, feedback on partially correct answers
to each question.}

\item{incorrect_feedback}{A string, feedback on incorrect answers to each question.}

\item{adapt_images}{A boolean, adapt the images so that they are a similar size.}

\item{width}{A integer, width of each image.}

\item{height}{A integer, height of each image.}
}
\value{
A \code{question_category} object.
}
\description{
Creates a \code{question_category} object.
}
\details{
Defines a category of questions to be included in the \emph{Moodle} question bank.
}
\examples{

qc <- question_category(category = 'Initial test')

}
\seealso{
Other question definition: 
\code{\link{define_question}()}
}
\concept{question definition}
