% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{read_question_csv}
\alias{read_question_csv}
\title{Read a question csv file}
\usage{
read_question_csv(file, sep = ",")
}
\arguments{
\item{file}{A string, name of a text file.}

\item{sep}{Column separator character.}
}
\value{
A data frame.
}
\description{
Reads a csv file of questions and returns a data frame.
}
\examples{

file <- system.file("extdata", "questions.csv", package = "moodef")
df <- read_question_csv(file = file)

}
\seealso{
Other support functions: 
\code{\link{create_question_csv}()},
\code{\link{create_question_data_frame}()},
\code{\link{vector_to_string}()}
}
\concept{support functions}
