% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simplified_types}
\alias{simplified_types}
\title{Simplified Question Types}
\format{
A character vector with 4 elements:
\describe{
\item{\code{""}}{Empty type}
\item{\code{"h"}}{Horizontal type}
\item{\code{"v"}}{Vertical type}
\item{\code{"x"}}{Other type}
}
}
\usage{
simplified_types
}
\description{
A vector of simplified question types used internally in the package.
}
\keyword{internal}
