% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{string_to_string_vector}
\alias{string_to_string_vector}
\title{Transform a String into a String-Formatted Vector}
\usage{
string_to_string_vector(str)
}
\arguments{
\item{str}{\code{character}
A string to transform.}
}
\value{
A string representing a vector in R syntax (e.g., \code{c("a", "b", "c")}), or \verb{"\\"\\""} if the input string is empty.
}
\description{
Converts a string into a string-formatted vector (e.g., \code{"a|b|c"} -> \code{c("a", "b", "c")}).
}
\keyword{internal}
