% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_xml.R
\name{xml_question}
\alias{xml_question}
\title{Define the question in xml}
\usage{
xml_question(type, name, questiontext, question_body)
}
\arguments{
\item{type}{A string, question type.}

\item{name}{A string, question name.}

\item{questiontext}{A string, question id number.}

\item{question_body}{A string, question id number.}
}
\value{
A string.
}
\description{
Define the question in xml
}
\keyword{internal}
