% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_xml.R
\name{xml_question_category}
\alias{xml_question_category}
\title{Define the question category in xml}
\usage{
xml_question_category(category)
}
\arguments{
\item{category}{A string, category name.}
}
\value{
A string.
}
\description{
Define the question category in xml
}
\keyword{internal}
