% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_xml.R
\name{xml_question_idnumber}
\alias{xml_question_idnumber}
\title{Define the question id number in xml}
\usage{
xml_question_idnumber(idnumber)
}
\arguments{
\item{idnumber}{A string, question id number.}
}
\value{
A string.
}
\description{
Define the question id number in xml
}
\keyword{internal}
