% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_xml.R
\name{xml_question_name}
\alias{xml_question_name}
\title{Define the question name in xml}
\usage{
xml_question_name(name)
}
\arguments{
\item{name}{A string, question name.}
}
\value{
A string.
}
\description{
Define the question name in xml
}
\keyword{internal}
