% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_xml.R
\name{xml_question_tags}
\alias{xml_question_tags}
\title{Define the question tags in xml}
\usage{
xml_question_tags(tag_values)
}
\arguments{
\item{tag_values}{A string, tag values.}
}
\value{
A string.
}
\description{
Define the question tags in xml
}
\keyword{internal}
