% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_xml.R
\name{xml_questiontext}
\alias{xml_questiontext}
\title{generate \code{questiontext} node}
\usage{
xml_questiontext(
  copyright,
  license,
  adapt_images,
  width,
  height,
  question,
  image,
  image_alt,
  type,
  author = "",
  fb_general = "",
  idnumber = ""
)
}
\arguments{
\item{copyright}{A string, copyright text to be included in each question that
is defined.}

\item{license}{A string, license text to be included in each question that is
defined.}

\item{adapt_images}{A boolean, adapt the images so that they are a similar size.}

\item{width}{A integer, width of each image.}

\item{height}{A integer, height of each image.}

\item{question}{A string, statement of the question.}

\item{image}{A string, optional, image file to include in the question.}

\item{image_alt}{A string, description of the image to include in the question.}

\item{type}{A string, question type.}

\item{author}{A string, author name to be included in each question that is defined.}

\item{fb_general}{A string, general feedback to be included in each question that is defined.}

\item{idnumber}{A string, idnumber to be included in each question that is defined.}
}
\value{
A string.
}
\description{
generate \code{questiontext} node
}
\keyword{internal}
