% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_category.R
\name{define_extended_question}
\alias{define_extended_question}
\alias{define_extended_question.question_category}
\title{Define an Extended Question in a Question Category}
\usage{
define_extended_question(
  qc,
  category,
  type,
  id,
  name,
  author,
  fb_general,
  fb_correct,
  fb_partially,
  fb_incorrect,
  question,
  image,
  image_alt,
  answer,
  a_1,
  a_2,
  a_3,
  a_4,
  a_5,
  a_6,
  fb_answer,
  fb_a_1,
  fb_a_2,
  fb_a_3,
  fb_a_4,
  fb_a_5,
  fb_a_6,
  tag_1,
  tag_2,
  tag_3,
  tag_4,
  tag_5,
  tag_6
)

\method{define_extended_question}{question_category}(
  qc,
  category = "",
  type = "",
  id = "",
  name = "",
  author = "",
  fb_general = "",
  fb_correct = "",
  fb_partially = "",
  fb_incorrect = "",
  question = "",
  image = "",
  image_alt = "",
  answer = "",
  a_1 = "",
  a_2 = "",
  a_3 = "",
  a_4 = "",
  a_5 = "",
  a_6 = "",
  fb_answer = "",
  fb_a_1 = "",
  fb_a_2 = "",
  fb_a_3 = "",
  fb_a_4 = "",
  fb_a_5 = "",
  fb_a_6 = "",
  tag_1 = "",
  tag_2 = "",
  tag_3 = "",
  tag_4 = "",
  tag_5 = "",
  tag_6 = ""
)
}
\arguments{
\item{qc}{A question category object. It should have a \code{questions} data frame
where new questions will be added.}

\item{category}{A character string specifying the category of the question.}

\item{type}{A character string indicating the type of the question (e.g., "multiple_choice").}

\item{id}{A unique identifier for the question.}

\item{name}{A character string representing the name of the question.}

\item{author}{The name of the author of the question.}

\item{fb_general}{General feedback for the question.}

\item{fb_correct}{Feedback displayed when the correct answer is selected.}

\item{fb_partially}{Feedback displayed for partially correct answers.}

\item{fb_incorrect}{Feedback displayed for incorrect answers.}

\item{question}{The text of the question.}

\item{image}{(Optional) Path to an image file associated with the question.}

\item{image_alt}{(Optional) Alternative text describing the image for accessibility.
Required if an image is provided.}

\item{answer}{The correct answer to the question.}

\item{a_1}{Additional possible answers for multiple-choice questions.}

\item{a_2}{Additional possible answers for multiple-choice questions.}

\item{a_3}{Additional possible answers for multiple-choice questions.}

\item{a_4}{Additional possible answers for multiple-choice questions.}

\item{a_5}{Additional possible answers for multiple-choice questions.}

\item{a_6}{Additional possible answers for multiple-choice questions.}

\item{fb_answer}{Feedback for the correct answer.}

\item{fb_a_1}{Feedback for additional answers.}

\item{fb_a_2}{Feedback for additional answers.}

\item{fb_a_3}{Feedback for additional answers.}

\item{fb_a_4}{Feedback for additional answers.}

\item{fb_a_5}{Feedback for additional answers.}

\item{fb_a_6}{Feedback for additional answers.}

\item{tag_1}{Tags to categorize the question.}

\item{tag_2}{Tags to categorize the question.}

\item{tag_3}{Tags to categorize the question.}

\item{tag_4}{Tags to categorize the question.}

\item{tag_5}{Tags to categorize the question.}

\item{tag_6}{Tags to categorize the question.}
}
\value{
Returns the updated question category object with the new question appended.
}
\description{
This function allows users to define an extended question, including metadata, feedback,
and optional image data, and append it to an existing question category (\code{qc}).
}
\details{
The function converts the parameters into a data frame and appends it to the
\code{questions} data frame of the provided question category object. If an image is included,
the \code{image_alt} parameter must also be defined to comply with accessibility standards.
}
\examples{

qc <- question_category(category = 'Initial test') |>
  define_extended_question(
    question = 'What are the basic arithmetic operations?',
    answer = 'Addition, subtraction, multiplication and division.',
    a_1 = 'Addition and subtraction.',
    a_2 = 'Addition, subtraction, multiplication, division and square root.'
  )

}
