% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_questions_extended.R
\name{extended_format_questions}
\alias{extended_format_questions}
\title{Format all questions in the data frame}
\usage{
extended_format_questions(qc)
}
\arguments{
\item{qc}{A \code{question_category} object.}
}
\value{
A string.
}
\description{
Format all questions in the data frame
}
\keyword{internal}
