% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_matching.R
\name{generate_matching}
\alias{generate_matching}
\title{Generate \code{matching} question}
\usage{
generate_matching(
  answer,
  n,
  rest,
  correct_feedback,
  partially_correct_feedback,
  incorrect_feedback
)
}
\arguments{
\item{answer}{A vector of strings.}

\item{n}{An integer, number or answers.}

\item{rest}{A vector, rest of answers.}

\item{correct_feedback}{A string.}

\item{partially_correct_feedback}{A string.}

\item{incorrect_feedback}{A string.}
}
\value{
A string.
}
\description{
Generate \code{matching} question
}
\keyword{internal}
