% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_multichoice.R
\name{generate_multichoice}
\alias{generate_multichoice}
\title{Generate \code{multichoice} question}
\usage{
generate_multichoice(answer, n, rest, correct_feedback, incorrect_feedback)
}
\arguments{
\item{answer}{A string.}

\item{n}{An integer, number or answers.}

\item{rest}{A vector, rest of answers.}

\item{correct_feedback}{A string.}

\item{incorrect_feedback}{A string.}
}
\value{
A string.
}
\description{
Generate \code{multichoice} question
}
\keyword{internal}
