% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_numerical.R
\name{generate_numerical}
\alias{generate_numerical}
\title{Generate \code{numerical} question}
\usage{
generate_numerical(answer, a_values, fb_answer, fb_a_values)
}
\arguments{
\item{answer}{A vector of strings.}

\item{a_values}{A vector, rest of answers.}

\item{fb_answer}{A string, answer feedback.}

\item{fb_a_values}{A vector, rest of answer feedback.}
}
\value{
A string.
}
\description{
Generate \code{numerical} question
}
\keyword{internal}
