% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_numerical.R
\name{generate_numerical}
\alias{generate_numerical}
\title{Generate \code{numerical} question}
\usage{
generate_numerical(answer, n, rest)
}
\arguments{
\item{answer}{A vector of strings.}

\item{n}{An integer, number or answers.}

\item{rest}{A vector, rest of answers.}
}
\value{
A string.
}
\description{
Generate \code{numerical} question
}
\keyword{internal}
