% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_xml.R
\name{generate_question}
\alias{generate_question}
\title{Generate question}
\usage{
generate_question(
  first_question_number,
  copyright,
  license,
  correct_feedback,
  partially_correct_feedback,
  incorrect_feedback,
  adapt_images,
  width,
  height,
  type,
  question,
  image,
  image_alt,
  answer,
  ...
)
}
\arguments{
\item{first_question_number}{An integer, first number to compose the question
names.}

\item{copyright}{A string, copyright text to be included in each question that
is defined.}

\item{license}{A string, license text to be included in each question that is
defined.}

\item{correct_feedback}{A string, feedback on correct answers to each question.}

\item{partially_correct_feedback}{A string, feedback on partially correct answers
to each question.}

\item{incorrect_feedback}{A string, feedback on incorrect answers to each question.}

\item{adapt_images}{A boolean, adapt the images so that they are a similar size.}

\item{width}{A integer, width of each image.}

\item{height}{A integer, height of each image.}

\item{type}{A string, question type (if needed).}

\item{question}{A string, statement of the question.}

\item{image}{A string, optional, image file to include in the question.}

\item{image_alt}{A string, description of the image to include in the question.}

\item{answer}{A string, correct answer to the question.}

\item{...}{A string, rest of the answers to the question.}
}
\value{
A string.
}
\description{
Generate question
}
\keyword{internal}
