% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_shortanswer.R
\name{generate_shortanswer}
\alias{generate_shortanswer}
\title{Generate \code{shortanswer} question}
\usage{
generate_shortanswer(answer)
}
\arguments{
\item{answer}{A string.}
}
\value{
A string.
}
\description{
Generate \code{shortanswer} question
}
\keyword{internal}
