% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_truefalse.R
\name{generate_truefalse}
\alias{generate_truefalse}
\title{Generate \code{truefalse} question}
\usage{
generate_truefalse(answer, fb_answer = "", fb_rest = NULL)
}
\arguments{
\item{answer}{A string.}

\item{fb_answer}{A string, answer feedback.}

\item{fb_rest}{A vector, rest of answer feedback.}
}
\value{
A string.
}
\description{
Generate \code{truefalse} question
}
\keyword{internal}
