% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_truefalse.R
\name{generate_truefalse}
\alias{generate_truefalse}
\title{Generate \code{truefalse} question}
\usage{
generate_truefalse(answer)
}
\arguments{
\item{answer}{A string.}
}
\value{
A string.
}
\description{
Generate \code{truefalse} question
}
\keyword{internal}
