% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_xml.R
\name{generate_xml}
\alias{generate_xml}
\alias{generate_xml.question_category}
\title{Generate the questions xml file}
\usage{
generate_xml(qc, file)

\method{generate_xml}{question_category}(qc, file = NULL)
}
\arguments{
\item{qc}{A \code{question_category} object.}

\item{file}{A string, file name.}
}
\value{
A \code{question_category}.
}
\description{
Generate the questions xml file
}
\seealso{
Other question definition: 
\code{\link{define_question}()},
\code{\link{question_category}()}
}
\concept{question definition}
