% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{is_numeric}
\alias{is_numeric}
\title{Check if it is numerica.}
\usage{
is_numeric(str)
}
\arguments{
\item{str}{A string.}
}
\value{
A boolean.
}
\description{
Check if it is numerica.
}
\keyword{internal}
