% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{string_to_string_vector}
\alias{string_to_string_vector}
\title{Transforms string into a vector in string format.}
\usage{
string_to_string_vector(str)
}
\arguments{
\item{str}{A string.}
}
\value{
A vector of strings.
}
\description{
Transforms string into a vector in string format.
}
\keyword{internal}
